fix: Consider queued updates for initial state
This solves issues when a component is mounted by
a state change, and which itself contains a hook on the
key that caused it to mount.

Because the correct value is still in the queue at mount time,
it won't be correctly set in the mounted component, and because
internal history updates don't trigger a state sync, the newly
mounted component won't re-render when the URL is updated.

Closes #359.