fix: Handle conflicts between RSC & local queue
When the server streams in a shallow: false navigation,
it may do so with an outdated querystring if there are
updates pending in the queue, and those would
overwrite the internal states.
So before triggering a sync, we merge the queue
onto the incoming search params and sync on the
result.