chore: Remove children pages to get CI to pass
What the actual freaking fuck?
The mere presence of those pages (in the app router),
and even if they only were to contain a static Hello world
output, would cause the **pages** router to fail to apply
updates, when there is a basePath, and on Next.js 13.5+.

Losing sanity here trying to figure out what's wrong,
and it's the middle of the night so I won't go deeper
into this right now.

See https://github.com/47ng/next-usequerystate/pull/397#issuecomment-1824889002