fix: Use pages router via internals
Going through the history API only on shallow updates doesn't
notify the pages router, which can handle shallow updates directly.

Definitely a hack, but it allows also setting the asPath for dynamic
routes. There is still an issue with basePath being applied twice,
but that's an internal Next.js bug for another day.