fix: Deprecate subscribeToQueryUpdates
It was broken in Next.js 14.0.3 due to shallow updates
not calling the patched history methods, and fixed in
Next.js 14.0.4.

However, since turning on the `windowHistorySupport` flag
causes `useSearchParams` to be reactive to shallow URL query
changes, there's no longer need for this method.

Marking it as deprecated now, and will be dropped along with
other deprecations in v2.0.0 once WHS lands in GA.

Closes #422.