fix: Use a traditional useEffect for emitter registration
useInsertionEffect was a hack to be notified from navigation events
in the same event loop tick as the Next.js router, but it causes issues
with tRPC's SSR and it looks like we could get by with a simple useEffect.

Testing this in CI to see if it's a viable fix, and there'll be internal docs
and scheduling code to update if all goes green.