fix: Don't rely on URL for initial state
This makes the first render on client side navigation
use the source URL rather than the destination,
which can cause issues if the first state is memoised
or used for some other purposes.

useSearchParams contains the destination values for searchParams,
however this might break in older versions of Next.js (TBC by CI).
In which case this fix might only land on v2.