chore: Use React 19 RC
There's now a peer dependency on the latest Next.js canaries onto React 19,
so until it is officially released, we'll have to use the `rc` dist tag.

Unfortunately, React 19 types don't export the `cache` experimental feature,
which Next.js embeds. So we have to cheat a bit with TypeScript here.

This fixes the canary test build. There are no public changes to deps in `nuqs`,
so intermediary releases until Next 15 / React 19 are published should be possible
without breaking semver in dependent apps.