fix: Deep equality check on cache parsing input
A change introduced in Next.js 15.0.0-canary.41 (precisely vercel/next.js#67105)
broke the referential stability that the cache.parse method was expecting to detect
`parse` being called with the same search params.

Now we're using a deep equality check by caching the serialized query string
that was parsed, and comparing it with the input when called again.