feat: Introducing adapters for other frameworks
BREAKING CHANGE: nuqs now requires wrapping your app
with a NuqsAdapter, which is a context provider connecting
your framework APIs to the hooks' internals.

BREAKING CHANGE: The `startTransition` option no longer
automatically sets `shallow: false`. The `Options` type
is no longer generic.

BREAKING CHANGE: The "use client" directive was not included
in the client import (`import {} from 'nuqs'`). It has now been added,
meaning that server-side code needs to import from `nuqs/server`
to avoid errors like:
```
Error: Attempted to call withDefault() from the server but withDefault is on
the client. It's not possible to invoke a client function from the server, it can
only be rendered as a Component or passed to props of a Client
Component.
```

Closes #603, #620.