chore: Run pkg.pr.new only if there are changes (#656)
* Fix: Run pkg.pr.new only if there are changes

* Update .github/workflows/pkg.pr.new.yml

Modified the check for changes to the nuqs package by adding an output variable skip=true when no changes are detected. This allows the workflow to continue without failing while enabling subsequent steps to conditionally execute based on this output. Additionally, adjusted the Git diff command to evaluate the entire commit range from base to HEAD, ensuring a more comprehensive check for any changes.

Co-authored-by: François Best <github@francoisbest.com>

* Update .github/workflows/pkg.pr.new.yml

Implemented a condition for the ‘Publish to pkg.pr.new’ step to execute only if changes were detected in the nuqs package. By adding if: steps.check-for-changes.outputs.skip != 'true', the workflow prevents unnecessary publishing actions when there are no changes, optimizing the deployment process and reducing potential errors.

Co-authored-by: François Best <github@francoisbest.com>

* Add newline before publish step for readability

This commit adds a newline between the 'Install dependencies' step and the 'Publish to pkg.pr.new' step in the workflow file to improve readability and maintainability of the code.

* Update .github/workflows/pkg.pr.new.yml

Co-authored-by: François Best <github@francoisbest.com>

* Update .github/workflows/pkg.pr.new.yml

Co-authored-by: François Best <github@francoisbest.com>

* Update pkg.pr.new.yml

Co-authored-by: François Best <github@francoisbest.com>

* Update pkg.pr.new.yml

Co-authored-by: François Best <github@francoisbest.com>

* chore: Update .github/workflows/pkg.pr.new.yml

---------

Co-authored-by: François Best <github@francoisbest.com>