chore: Install Next.js versions to both e2e & nuqs
This solves some incompatibilities when running tests related to the React version
embedded by Next and exported by nuqs.

Symptom was that all pages in the pages router crashed on mount.

Supersedes #677, which attempted to use the catalogs feature of pnpm to keep things in sync,
but isn't practical to inject custom versions in a CI context.