feat: Add experimental async cache
To leverage Next.js 15's async searchParams prop and PPR,
we can defer the parsing of search params to where
they are actually used, rather than blocking rendering for
top-level elements of the page.

Highly experimental, not sure how best to introduce it retro-compatibly
with the existing cache (with sync/async overloads, but still top-level
blocking PPR). So for now it's a fork.