test: Add rate limit stress test
This tests how the throttled queue alleviates browsers rate-limiting
calls to the History API for URL updates.

Updates to a counter are made every 20ms for one minute, and
the test keeps count of the result for each URL update (the resolved
Promise from the state updater function).

Note: updates are not made using the internal state of useQueryState,
as it's data-racing against the URL. Instead, we keep a steady
increasing count and only render the current URL value for progress.