test: Add stitching/sequencing test
Fixes a bug in the React adapter which wasn't using the history
update marker. Wasn't really an issue since there was no
side-effect before, but it got detected by that stitching test
when the queues got flushed on the first update.

Fixes a bug in the Next.js pages adapter where the navigation detection
could not dissociate nuqs updates (where we don't want to clear the queues)
from external navigations. Using a poor man's mutex to disable queue clearing
while nuqs updates are in flight.