fix: Move throttle queue reset responsibility to adapters
Each will behave differently in terms of concurrent render sequencing,
so we elected repro-702 to a shared e2e test to verify that updates
that mount component ensure their states are correct wrt/ queued updates.
Each adapter will reset the throttle queue when they set their optimistic
useSearchParams, which can happen at different moments of the update
process (eg: the app router needs to do this in a startTransition for useOptimistic).

This changes the behaviour of the internal debounce & throttle queues,
so tests were updated accordingly.
There is still a lot of coupling between all these internal data structures,
but if tests are all green in CI, that'll be a step closer towards refactoring
with a peace of mind.