ref: Allow adpaters choosing the queue reset behaviour
Some adapters will need to reset the throttle queue in different
ways, to ensure a stable number of renders and initialising mounted
components with queued states in the right way.

By default, the queue is reset as it was before, and adapters can opt-out
of this auto-reset and have the responsibility of doing it when a URL
udpate occurs.