feat: Partial output for TSR support
TanStack Router handles defaults a bit differently
than nuqs, and would require a Partial output type
for Links not to have to specify every single search
param value for a given route.

Moreover, the output of validateSearch is reflected in
the URL a page load time, which contradicts the "clean URL first"
approach taken by nuqs (it also reflects null values which is not
desirable).