fix: reset throttle queue before first push
This solves a few issues:
- The reset occurs synchronously with push,
so in the same event loop tick where the internal state
is updated: both states are invalidated at the same time
- The queue doesn't build up over time
- No more hacks in the pages router 🙌

We still need the auto-reset mechanism for the testing adapter,
although I feel like we could probably get rid of that too.
One thing at a time.