doc: revert cacheComponents
While the idea of pre-rendering the outer shell makes sense for dynamic pages,
the fact that it's no longer possible to obtain purely static outputs
(because of the "use cache" invalidation mechanism needing a lambda to run)
is kind of a deal-breaker for highly static content, where we occasionnaly use ISR
to refresh parts of the page that depend on external APIs.

Might revisit in the future.